enable lints

+++ spam/vote_spam_policy.go
@@ -28,11 +28,11 @@ func (b *blockRejectInfo) add(rejected bool) {
 
 var maxMinVotingTokens, _ = num.UintFromString("1600000000000000000000", 10)
 var (
-	//ErrPartyIsBannedFromVoting is returned when the party is banned from voting
+	// ErrPartyIsBannedFromVoting is returned when the party is banned from voting.
 	ErrPartyIsBannedFromVoting = errors.New("party is banned from submitting votes in the current epoch")
-	//ErrInsufficientTokensForVoting is returned when the party has insufficient tokens for voting
+	// ErrInsufficientTokensForVoting is returned when the party has insufficient tokens for voting.
 	ErrInsufficientTokensForVoting = errors.New("party has insufficient tokens to submit votes in this epoch")
-	//ErrTooManyVotes is returned when the party has voted already the maximum allowed votes per proposal per epoch
+	// ErrTooManyVotes is returned when the party has voted already the maximum allowed votes per proposal per epoch.
 	ErrTooManyVotes = errors.New("party has already voted the maximum number of times per proposal per epoch")
 )
 
@@ -59,7 +59,7 @@ type VoteSpamPolicy struct {
 	lock                    sync.RWMutex                                     // global lock to sync calls from multiple tendermint threads
 }
 
-//NewVoteSpamPolicy instantiates vote spam policy
+// NewVoteSpamPolicy instantiates vote spam policy.
 func NewVoteSpamPolicy(minTokensParamName string, maxAllowedParamName string, log *logging.Logger) *VoteSpamPolicy {
 	return &VoteSpamPolicy{
 		log:                   log,
@@ -188,12 +188,12 @@ func (vsp *VoteSpamPolicy) Deserialise(p *types.Payload) error {
 	return nil
 }
 
-//UpdateUintParam is called to update Uint net params for the policy
-//Specifically the min tokens required for voting
+// UpdateUintParam is called to update Uint net params for the policy
+// Specifically the min tokens required for voting.
 func (vsp *VoteSpamPolicy) UpdateUintParam(name string, value *num.Uint) error {
 	if name == vsp.minTokensParamName {
 		vsp.minVotingTokens = value.Clone()
-		//NB: this means that if during the epoch the min tokens changes externally
+		// NB: this means that if during the epoch the min tokens changes externally
 		// and we already have a factor on it, the factor will be applied on the new value for the duration of the epoch
 		vsp.effectiveMinTokens = num.Zero().Mul(vsp.minVotingTokens, vsp.minVotingTokensFactor)
 	} else {
@@ -202,8 +202,8 @@ func (vsp *VoteSpamPolicy) UpdateUintParam(name string, value *num.Uint) error {
 	return nil
 }
 
-//UpdateIntParam is called to update iint net params for the policy
-//Specifically the number of votes to a proposal a party can submit in an epoch
+// UpdateIntParam is called to update iint net params for the policy
+// Specifically the number of votes to a proposal a party can submit in an epoch.
 func (vsp *VoteSpamPolicy) UpdateIntParam(name string, value int64) error {
 	if name == vsp.maxAllowedParamName {
 		vsp.numVotes = uint64(value)
@@ -213,7 +213,7 @@ func (vsp *VoteSpamPolicy) UpdateIntParam(name string, value int64) error {
 	return nil
 }
 
-//Reset is called at the beginning of an epoch to reset the settings for the epoch
+// Reset is called at the beginning of an epoch to reset the settings for the epoch.
 func (vsp *VoteSpamPolicy) Reset(epoch types.Epoch, tokenBalances map[string]*num.Uint) {
 	vsp.lock.Lock()
 	defer vsp.lock.Unlock()
@@ -259,7 +259,7 @@ func (vsp *VoteSpamPolicy) Reset(epoch types.Epoch, tokenBalances map[string]*nu
 	}
 }
 
-//EndOfBlock is called at the end of the block to allow updating of the state for the next block
+// EndOfBlock is called at the end of the block to allow updating of the state for the next block.
 func (vsp *VoteSpamPolicy) EndOfBlock(blockHeight uint64) {
 	vsp.lock.Lock()
 	defer vsp.lock.Unlock()
@@ -300,7 +300,7 @@ func (vsp *VoteSpamPolicy) EndOfBlock(blockHeight uint64) {
 	}
 
 	// check if we need to increase the limits, i.e. if we're below the max and we've not increased in the last n blocks
-	if (vsp.lastIncreaseBlock == 0 || blockHeight > vsp.lastIncreaseBlock+uint64(numberOfBlocksForIncreaseCheck)) && num.Zero().Mul(vsp.minVotingTokens, vsp.minVotingTokensFactor).LT(maxMinVotingTokens) {
+	if (vsp.lastIncreaseBlock == 0 || blockHeight > vsp.lastIncreaseBlock+numberOfBlocksForIncreaseCheck) && num.Zero().Mul(vsp.minVotingTokens, vsp.minVotingTokensFactor).LT(maxMinVotingTokens) {
 		average := vsp.calcRejectAverage()
 		if average > rejectRatioForIncrease {
 			vsp.lastIncreaseBlock = blockHeight
@@ -310,7 +310,7 @@ func (vsp *VoteSpamPolicy) EndOfBlock(blockHeight uint64) {
 	}
 }
 
-// calculate the mean rejection rate in the last <numberOfBlocksForIncreaseCheck>
+// calculate the mean rejection rate in the last <numberOfBlocksForIncreaseCheck>.
 func (vsp *VoteSpamPolicy) calcRejectAverage() float64 {
 	var total uint64 = 0
 	var rejected uint64 = 0
@@ -324,8 +324,8 @@ func (vsp *VoteSpamPolicy) calcRejectAverage() float64 {
 	return float64(rejected) / float64(total)
 }
 
-//PostBlockAccept checks if votes that made it to the block should be rejected based on the number of votes preceding the block + votes seen in the block
-//NB: this is called as part of the processing of the block
+// PostBlockAccept checks if votes that made it to the block should be rejected based on the number of votes preceding the block + votes seen in the block
+// NB: this is called as part of the processing of the block.
 func (vsp *VoteSpamPolicy) PostBlockAccept(tx abci.Tx) (bool, error) {
 	party := tx.Party()
 
@@ -387,12 +387,11 @@ func (vsp *VoteSpamPolicy) PostBlockAccept(tx abci.Tx) (bool, error) {
 	}
 	vsp.blockPostRejects.add(false)
 	return true, nil
-
 }
 
-//PreBlockAccept checks if the vote should be rejected as spam or not based on the number of votes in current epoch's preceding blocks and the number of tokens
-//held by the party.
-//NB: this is done at mempool before adding to block
+// PreBlockAccept checks if the vote should be rejected as spam or not based on the number of votes in current epoch's preceding blocks and the number of tokens
+// held by the party.
+// NB: this is done at mempool before adding to block.
 func (vsp *VoteSpamPolicy) PreBlockAccept(tx abci.Tx) (bool, error) {
 	party := tx.Party()
